Add BDF path attribute to the status struct
The Bus-Device-Function path is another way of representing the PCI
address of a device. Instead of using the full BDF of the device,
use the BDF of a base device followed by a path of device and functions.

This is similar to the "device scope" structure in the Intel VT-d
spec, Section 8.3.1. It will also be useful for specifying the
switch ports on the command line to disable ACS in the system.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>