Fix build regression in recent gui_main() change
There are two gui_main() functions. When libcurses is not available
(ie. when cross compling for windows), a stub function is used.
A recent change to the gui_main() prototype forgot to make the
analagous change to the stub function and therefore caused
a build regression.

While we are at it fix, the function prototype in gui.h which should have
proper argument names.

This was caught by travis on the latest devel branch.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>
Fixes: 50a6136121e5 ("Add new option --bw_type for command gui and...")
Signed-off-by: Logan Gunthorpe <logang@deltatee.com>