if errno not equal to 0, keep its value instead of set to ENODEV
In switchtec_open funcion, if no device was found, while errno equal
to 0, this will introduce false "success" prompt, so in this case,
set the errno to ENODEV unconditionly.
But if errno already set by system api, its value should be kept
instead of set to ENODEV.

reason for this patch:
for uart platform, we use flock with unblock
if flock return on error when the file is locked, errno is set to
"EWOULDBLOCK", and perror message output is "Resource temporarily
unavailable", which is the expected.
But if set errno to "ENODEV", the message output is "No such device",
which confuse user.