Map the entire gas space twice for two different section
Previously, the whole gas sapce be mapped only once in user tool,
it work with driver which set the whole gas memory to UC- in PAT.
But it does not work with driver which set WC for first 4K MRPC
region, while keep the left same as before, that's UC-.
Map twice fix this issue.

To fix this, we reserve a virtual address space for the entire GAS region
then use two MAP_FIXED calls to allocate the two different regions.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>