Fix logical error when parse partition and port of event
In event-wait function, there should be negtive value for port
when event type is SWITCHTEC_EVT_PART, which means port cannot
be specified for this event type.

Previsouly, the default port value (-1) will lead to error
return of this function.
Fix it by error return only when port is set a non-negtive
value.

Signed-off-by: Wesley Sheng <wesley.sheng@microchip.com>