Fix segmentation fault case for UART interface on 32bit system
The ap(type of va_list) is undefined after return of function
vsnprintf that uses va_arg(ap, type).
Fix the undefined case by rearaange the sequence of variable
arguments for send_cmd function, write_data and write_crc are
followed when write_bytes is non zero, they can retrieve by
va_arg. For the generic arguments used by both gas read and
write, are passed to vsnprintf.

Root cause for segmentation fault:
The va_arg after vsnprintf in send_cmd function, return the
value of the first variable argument after the fixed arguments,
instead of the expected second one.
Because the first varialbe arguments is an offset inf GAS, and
the second variable arguments is a memory address; take the
offset(relative low) as memory address and deferencing cause
the segmentation fault.
This issue found on i686, while not on x86_64.

References:
[1] https://linux.die.net/man/3/vsnprintf
The functions vprintf(), vfprintf(), vsprintf(), vsnprintf()
are equivalent to the functions printf(), fprintf(), sprintf(),
snprintf(), respectively, except that they are called with a
va_list instead of a variable number of arguments. These
functions do not call the va_end macro. Because they invoke
the va_arg macro, the value of ap is undefined after the call.
See stdarg(3).
[2] https://linux.die.net/man/3/stdarg
If ap is passed to a function that uses va_arg(ap,type) then
the value of ap is undefined after the return of that function.

Signed-off-by: Wesley Sheng <wesley.sheng@microchip.com>