Fix output of event-wait command with false info
The output of event-wait command with false info when timeout
happened.

The root cause is: the data structure struct switchtec_event_summary
sum on stack doesn't zeroize, and in timeout situation, no update of
this data structure, then the output is random and false.

Fix it by zeroize the structure before use it.

The bug is found when test on Windows platform with command:
switchtec.exe event-wait switchtec0 -t 1000

Signed-off-by: Wesley Sheng <wesley.sheng@microchip.com>