Enhance switchtec error handling to make it MRPC aware
In the switchtec API functions, errno was used to store either MRPC
specific errors or other systme errors. To distinguish between these two
cases, we use the 31st bit (second highest bit) of errno to indicates
it's a MRPC error (bit set) or a system error (bit clear).

A new global variable mprc_error_cmd was also introduced to store
the MRPC command ID when the errno was for MRPC, as the firmware
don't ensure all the errno are unique across all MRPC commands.