Fix cfg_positive_handler function bug
cfg_positive_handler function was designed for check and parse
the positive argument, and warning when negative or zero argument
was input. But current implementation negative and zero argument
input was not filter out, because strtoul was used.
Fix it by change from strtoul to strtol, and doulbe check if the
output of strtol is negative or zero, warning accordingly.

Signed-off-by: Wesley Sheng <wesley.sheng@microchip.com>