Check RSA_get0_key() presense in libssl-dev
Libssl-dev 1.0 or older versions don't have this function. Add
presense check of this function in ./configure and define it if
it is not present in the installed version of libssl-dev.

Note: We still need to handle the case when libssl-dev is not available
in upcoming commit. We should allow user who doesn't care about
'recovery' features to compile the code.