Fix a bug of fw-read on gen4 switch
The fw image layout on the fw partition changed in Gen4, the metadata is
placed at the very beginning of the partition, rather than at the foot
in Gen3. The Gen4 fw image file layout is also changed, the image body
is placed at offset 0x1000, rather than immediatly after the image header.

This fix makes sure for a fw-read on Gen4 switch that
1, The image body is read from correct offset from a fw partition.
2, The image body is written to the correct offset of a image file.