Fix gas read / write in Windows
gas read / write commands fail in Windows either silently or with error messages. This is because "CFG_SIZE_SUFFIX" is used as the argument type for address / count arguments, which parses them as 64-bit ssize_t types, but the variables are declared as 32-bit longs. The argument values are then parsed incorrectly.

The fix is to change the "count" variables to size_t types, and change address argument types back to CFG_LONG_SUFFIX as GAS addresses are always 32 bits.