Fix evcntr-wait timeout segfault in Windows
The evcntr-wait command will segfault in Windows when the timeout expires. This is because switchtec_evcntr_wait() passes NULL to switchtec_event_wait_for()'s "res" parameter, which makes its way to gasop_event_summary(), which then tries to memset it.

This is fixed by adding a NULL parameter check to gasop_event_summary(). The equivalent function for linux (linux_event_summary()) already has a NULL check, but should be updated to return 0 instead of an error code, as the error code prevents the timeout message from being printed.