Fix the errno of switchtec_fw_part_info
In fw_toggle(), after the toggle operation, it will get fw info and
display. In the case of a failed toggle operation, it is observed that
the fw-toggle command still show suceess after the fw info display.

The reason is that when get fw info, the switchtec_fw_part_info(), which
is called when getting the fw info, will clear the errno in some cases
when it ends up succeeded. Normally, we don't update the errno on
success, remove the errno clear code to make sure it's untouched on
success return of the function.

In addtiion, a following patch will fix the error check in fw_toggle.