Fix fw_toggle() error info on failure
In fw_toggle(), after the toggle operation, it will get fw info and
display. In the case of a failed toggle operation, it is observed that
the fw-toggle command still show suceess after the fw info display.

The reason is that the errno might be updated between the toggle
function and switchtec_perror(), so the errno processed may not the
one from toggle function.

As we do want to show the success or failure info at the of the output,
therefore we use a temporary var to hold it after toggle and pass it to
switchtec_perror() at the end.