Command description updates
Various command descriptions, command argument descriptions, and other text outputs are incorrect, have spelling/grammar errors, or are inconsistent. Also, command descriptions are duplicated in both the commands[] arrays and the command functions themselves (for passing the description to argconfig_parse()).

The following changes are being applied to the main, GAS, fabric, and mfg commands:
-statically define all command descriptions and use these for the commands[] array and argconfig_parse() calls
-update command and argument descriptions as appropriate for correctness, spelling/grammar, and consistency

The following changes are also being made:
-fix some argv[] usages in main.c error outputs
-fix a few other spelling/grammar issues throughout the code