Fix 'fw-toggle' error in BL2 boot phase.
Command 'fw-toggle' in BL2 boot phase prints error from function 'switchtec_perror'. This is due to function 'switchtec_fw_toggle_active_partition' set errno to non-zero even when it succeeds. Since the value in errno is significant only when the return value of the call indicated an error, we should check ret code before saving the value of errno.