Use 'gas_write32_no_retry' for specific MRPC commands.
Due to the unreliable nature of hardware communication, function __gas_write32() is implemented with automatic retry. This poses a potential issue when a command is critical and is expected to be sent only once (e.g., command that adds a KMSK entry to chip OTP memory). Retrying could cause the command be sent multiple times (and multiple KMSK entry being added, if unlucky). We filter out specific commands and use 'no retry' version of 'gas_write32' for these commands.