Improve switchtec_get_fw_version() to support remote PAX
It is observed that when running fw-info on a remote PAX, the fw version
shown is the one of the local PAX. The reason is that the
swtichtec_get_fw_version() API used is implemented by memory mapped GAS
access (always to local GAS). Therefor, it's unaware of PAX ID.

Fix it by implementing the swtichtec_get_fw_version() by leveraging
switchtec_fw_part_summary() which call MRPC (PAX ID aware) to get the version.

Note: the fix is generally for Gen4 PAX, as we don't support Gen3 PAX.