Fix "make install" on MSYS2
"make install" fails in an MSYS2 environment, complaining that it can't create the symlink "/usr/local/lib/libswitchtec.dll.a". This is because the source of the symlink doesn't exist, and since MSYS2 just creates copies instead of symlinks, the copy fails.

Additionally, the install/uninstall Makefile targets run ldconfig commands. $LDCONFIG isn't defined in MSYS2 as ldconfig isn't supported, so these commands fail.

The symlink creation is now only done in a non-MinGW environment, as are the ldconfig commands.