Fix "fw-info" command for Gen3 devices on Windows
Running the "fw-info" command in Windows for a Gen3 device gives the error "A device attached to the system is not functioning." This is because the MRPC_GET_DEV_INFO command is being issued but the command isn't supported by Gen3 firmware. switchtec_get_device_info() attempts to account for this by checking for ERR_MPRC_UNSUPPORTED, but the MRPC return value isn't available when the MRPC status indicates failure.

This is fixed by calling switchtec_boot_phase() from fw_info to get the boot phase rather than using switchtec_get_device_info().