Add switchtec_nvme_admin_passthru function
This function sends an Admin Command, as defined in Chapter 5 of NVMe specification v1.4, to the device's Admin Submission Queue, and reads back reply from the device.

As the reply from the device is 4096 bytes long, exceeding the maximum data length of an MRPC packet, the command is further divided into 3 subcommands: sending the Admin Command, reading the reply data (in chunks), and completing the command.