cli: Add the pattern command
The pattern command can enable or disable the pattern generator and/or
the monitor.

Pass the -d flag to disable both. And either or both of the -g and -m
flags to enable the generator and monitor for a port respectively.

Running with no flags prints the current state. If the monitor is
enabled it will print the error counts for all the lanes.