lib: Introduce function to calculate lane masks
The lane ID in the switch is usually calcualted as twice the physical
port ID. However, there are some exceptions with ports above 48.

Introduce switchtec_calc_lane_mask() to calculate a mask of lane ids
from a physical port number, lane id (within the port) and number
of consecutive lanes.