Add github-actions CI definition
TravisCI switched to a new billng model back in November which means
that free plans only have a fixed non-renewing amount of credit. OSS
projects can apply for free credits but this is a hassle and there's no
guarantee they'll approve a recurring amount.

Instead switch to using github-actions which is free for public repos.

So add a github actions yml file that mimics the actions previously done
by Travis.

Note, there are some changes: specifying -Werror everywhere was
redundant with the W=1 in the make line; and using the version of
sparse in Ubuntu is now preferred over building from scratch.