cli: Introduce stack-bif command
The stack-bif command allows the user to get and dynamically change the
bifurcation of a given stack.

Bifurcations are specified on the command line  with a series of numbers
(0, 1, 2, 4, 8, 16). Each number specifies the width of the next
available port (aligned to it's appropriate width). A zero specifies
a given port should be disabled.

Some examples:

This command will disable the first 4 ports and create a x8 out of the
last half of the stack:

  switchtec stack-bif <device> -s <stack_id> 0 8

This command will create a x8, port 4 will be disabled, 5 will be an
x2 and 6 will be an x4:

  switchtec stack-bif <device> -s <stack_id> 8 0 2 4