Add Gen5 firmware image header type support
With the introduction of RIOT image, the Gen5 firmware image
format has diverged from that of Gen4.
To differentiate, a new 'fw_id' field is added in the image
header. A non-zero 'fw_id' means the image is for Gen5 device.
New functions hdr_type2_id_gen4/5() are defined so that the
firmware type field can be mapped correctly.