lib: Correct and use helper to determine max number of ports
The issue in #317 was caused by accessing garbage data in ports beyond
the supported max for a device.

According to commit 2bf2c880fb2b ("lib: Increase SWITCHTEC_MAX_PORTS to
60") and commit 7e0d165b1aaf ("Fix segmentation fault in 'status'
command"), gen5 supports 60 ports and gen4 supports 52 ports. The fall
back value (added in this change) is 48 as it originally was in
2bf2c880fb2b.

Fixes #317