lib/platform/linux: Ensure entire switchtec_event_summary struct is set
In normal cases, event_summary_copy() will set the entire
switchtec_event_summary struct. However, if the library falls back to
the smaller SWITCHTEC_IOCTL_EVENT_SUMMARY_LEGACY, then a size less than
SWITCHTEC_MAX_PFF_CSR will be passed to the function and a portion
of the function will be left unset.

There are a number of places in the code where this function gets
called pointing to an unitialized structure on the stack. So fix
this issue by ensuring event_summary_copy() sets the entire
structure.

Resolves: #325 ("events command returning 'Invalid argument'")