cli: Fix an 'events -r' issue
It was observed that the option '-r' doesn't take effect unless the
option '-a' is specified at the same time. The option '-a' should only
control the showing of events and has no control for clearing. Fix this
to allow option '-r' clear all events no matter if option '-a' is
specified.