Add support for Gen5 Port Eq Dump commands
Add support for Gen 5 for existing port eq commands.

Port Equalization commands supported with the changes
1. port-eq-txcoeff
2. port-eq-txfslf
3. port-eq-txtable

Change port eq txfslf to print the width of the current port instead of
the negotiated link width to match tx coeff command for gen5. This
change is to match the MRPC gen5 spec.
(The -P (previous) equalization values are not currently supported in
the Gen5 version due to firmware change requirement.)