Added tlp injection command
TLP injection command added to use the defined MRPC interface
to send a raw TLP to a given destination port. User defined type
and enable the ecrc.

Raw tlp data is accepted as a string of characters and converted to
appropriate dword type and sent as part of new structure defined by
spec. Up to 132 dwords of raw tlp data can be sent a time, this limit is
defined by a macro in the switchtec header file.