Add ordered set commands
Add requested OSA ordered set analyzer coammnds to the switchtec user
commands to handle OSA split into 7 commands which reflect each MRPC
sub command as part of the OSA group.

osa - command which triggers specific osa operations on a given stack

osa config type - configure the osa type 4 different masks. Lane mask
to trigger on specific lanes, a directional mask, a link rate mask, and
ordered set type mask, each on a given stack

osa config pattern - configure osa pattern with similar masks as
previous minus a directional mask. Value and mask dwords are added to
configure the value of the pattern and the mask for the pattern.
The corresponding bit in the pattern value is used for matching.

osa config misc - This is for triggering the configuration on trigger
types not covered by the previous two commands. There a trigger enable
mask for 3 corresponding types: LTMON, other HW blocks, general
purpose.

osa capture control - This command allows the user to control how the
OSA is captured along lane and direction using masks as well how many
entries to capture, what types, and specific modes to capture on.

osa dump config - This command will dump the config on the specified
stack.

osa dump data - This command will dump the data for the osa given a
lane mask and directional mask, for a specific stack.