Update fw version string to expected size
changed the 64 byte buffer sent to switchtec_get_fw_version
function to 32. Since we are expecting the version information
to be at most 32 bytes there is no need to have the destination
buffer set to 64 when strncpy is called in the
switchtec_get_fw_version function within lib/platform/platform.c.