update loopback to gen5 + update help message
Update the loopback code to add a new function to handle
gen 5 loopback seperatly due to the changes compared to gen4.
Renamed types for gen5 to better reflect spec.
Added checking for ensuring there is no crossover between the
gen 4 / 5 arguements.
Added check for only gen4 args on a gen5 system.
Added arguements for the parallel option and the external option
and added additional information in the help option to show the
arguement supports gen 4 / gen 5.