add command option for riotcore
Add the requested feature to toggle the riotcore partition option, add
as part of the existing fw-toggle command. As part of these changes a
new 8bit field to toggle the riotcore is included to the toggle MRPC
cmd. FW partition line prints RIOT as inactive or active in sections
of the fw_info command and fw_toggle command where approriate.